Lab 1 for real
So I realized I'm a **Big Stupid** and uploaded just a LaTeX file with some regex because I thought you wanted us to just show that we knew how to use LaTex . . . 
Anyways, here is Lab 1. It's late as heck but I'm going to do them all instead of skipping them like an idjit.