Ok for realz
I hate myself a bit, and its 5.
Ok so this time around I implemented a scanner for taking in a file on your computer, I'll slap that in the readme in a sec. Next up, I added warnings (for the $ bit). I'm hoping this is the last commit for this project. Now time to sleep.