Big Brain
Trees in Java are dumb. You either get an auto-balancing tree, or one that can only have two kids. Like what? So I came up with a solution! Screw trees. Like Alan says, recursion is probably the answer and the question this time is how to make a List work like a tree.