Making the bones
I'm at work so not a tremondous amount of code here, but this should be *most* of the bones for the code. I'm probably going to need to come back to this to add more methods or change how these function but for now I'm going to try and flesh these out and see how far I can get.