Tables no longer break (TM)
well. Tables and values huh? Turns out storing the whole value equation in tables *may* have some less than desirable effects on how the table looks especially when the value field is smack dab in the middle of the table. so uh. changed the order and now it maybe hopefully won't break/look a mess.